feat alter credit table from int to decimal
task 7 solved. changed to decimal and ran dotnet ef migrations add V2__ModifyCourseCredits --context StudentManagementContext --output-dir Migrations/StudentManagement

+++ Models/Course.cs
@@ -9,7 +9,7 @@ public partial class Course
 
     public string Title { get; set; } = null!;
 
-    public int Credits { get; set; }
+    public decimal Credits { get; set; }
 
     public int InstructorId { get; set; }
